// src/app/models/Experience.ts
export interface Experience {
  /** First day in the company  */
  from: String;
  /** Last day in the company */
  to: String;
  /** Name of the job */
  job: String;
  /** Name of the company */
  company: String;
  /** List of the highlights. Things you did, achievements... */
  highlights: String[];
}